Watson_Tucker_Capstone1/requirements.txt
alembic==0.9.9
backcall==0.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.15.0
chardet==3.0.4
click==6.7
colorama==0.4.4
cryptography==36.0.1
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
Flask==1.0.2
Flask-Dance==0.14.0
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-OAuth==0.12
Flask-OAuthlib==0.9.4
Flask-SQLAlchemy==2.3.2
Flask-WTF==1.0.0
httplib2==0.11.3
idna==2.6
install==1.3.5
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.17.2
Jinja2==2.10
jupyter-client==7.1.2
jupyter-core==4.9.2
lazy==1.3
Mako==1.0.7
MarkupSafe==1.1.0
mysql-connector==2.2.9
mysql-connector-python==8.0.17
mysql-connector-python-rf==2.2.2
nest-asyncio==1.5.4
numpy==1.19.5
oauth2==1.9.0.post1
oauthlib==2.0.7
openpyxl==3.0.9
pandas==1.1.5
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.28
protobuf==3.19.4
pycparser==2.21
Pygments==2.11.2
PyMySQL==1.0.2
python-dateutil==2.8.2
python-editor==1.0.3
pytz==2021.3
pywin32==303
pyzmq==22.3.0
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
SQLAlchemy==1.2.6
SQLAlchemy-Utils==0.33.2
tornado==6.1
traitlets==4.3.3
urllib3==1.22
URLObject==2.4.3
wcwidth==0.2.5
Werkzeug==0.14.1
wincertstore==0.2
WTForms==2.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
